pyproject.toml
[project]
name = "ai-frontier"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "feedparser>=6.0.12",
    "html-to-markdown>=2.7.1",
    "markdown>=3.7.0",
    "google-api-python-client>=2.150.0",
    "google-genai>=1.0.0",
    "psycopg2-binary>=2.9.11",
    "pydantic>=2.0.0",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "sqlalchemy>=2.0.44",
    "youtube-transcript-api>=1.2.3",
]

[dependency-groups]
dev = [
    "ipykernel>=7.1.0",
]
